# encoding/json: optimize field sorting with slices and cmp